lambda/layers/requirements.txt
python-jose==3.3.0
requests==2.31.0
requests-toolbelt==1.0.0
urllib3<=2.0.0